practic_02/django/requirements.txt
Django==4.0.4; python_version >= '3.8'
Django==3.2.14; python_version < '3.8'
djangorestframework==3.12.4
django-storages[google]==1.13
django-environ==0.9.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
google-cloud-secret-manager==2.12.0
drf-yasg